Move useHelpdeskAccess() custom hook out of app component and into utilities file
